Struct slr_parser::Lexer

pub struct Lexer<'l, 's> where
    's: 'l,  {
    pub cur_token: Option<Result<Token<'s>, Error>>,
    pub next_token: Option<Result<Token<'s>, Error>>,
    // some fields omitted
}

A type handling the lexing.

Fields
cur_token: Option<Result<Token<'s>, Error>> next_token: Option<Result<Token<'s>, Error>>
Methods

impl<'l, 's> Lexer<'l, 's>

pub fn new(source: &'l mut Source<'s>) -> Lexer<'l, 's>

Creates a new lexer from a source. The source will be reset by this operation, and must not be used with any spans created from a previous lexing done with that source.

Important traits for Source<'l>
impl<'l> Iterator for Source<'l> type Item = char;

pub fn get_source(&self) -> &Source<'s>

pub fn next(&mut self) -> Option<Result<Token<'s>, Error>>